swc-installation-test-2.py: Allow multiple URLs for installation help
So we can link to both SWC and the upstream project.

I added an extra layer of indentation for the urls to set them off
from the rest of the instructions.  We want to make it easy for folks
to focus on the text, while the URLs are just for copy/pasting.