swc-installation-test-2.py: Targeted links to workshop setup instructions
With [1], we can mostly point folks to the right place if we can't
find a particular package.

There's some funkiness around editors, since most of our editor
instructions are just links to the upstream projects, and we have
different suggestions for each operating system.

We're also pointing most users at Anaconda, instead of listing the
packages we need explicitly [2,3,4], it's hard to know which specific
PythonPackageDependencies should also link to #python-{system}.  I've
left them off entirely for now, but instructors who want to link to
their setup instructions when we can get a Python version but not, for
example, a NumPy version should add a ('numpy', 'python') entry to the
check/check_id loop.

[1]: https://github.com/swcarpentry/workshop-template/pull/171
[2]: https://github.com/swcarpentry/bc/issues/724#issuecomment-56095114
[3]: https://github.com/swcarpentry/bc/issues/724#issuecomment-56095877
[4]: https://github.com/swcarpentry/bc/issues/724#issuecomment-56096123